requirements.txt
absl-py==2.1.0
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
annoy==1.17.3
anyio==3.7.1
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
backoff==2.2.1
beautifulsoup4==4.12.2
bleach==6.1.0
blis==0.7.11
boto3==1.34.37
botocore==1.34.37
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cohere==4.34
comm==0.2.1
confection==0.1.4
contourpy==1.2.0
cycler==0.10.0
cymem==2.0.8
dataclasses-json==0.6.3
datasets==2.14.6
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
distro==1.9.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
executing==2.0.1
fastapi==0.103.1
fastavro==1.8.2
fastjsonschema==2.19.1
filelock==3.13.1
fire==0.5.0
fonttools==4.47.2
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.10.0
fst-pso==1.8.1
FuzzyTM==2.0.5
gensim==4.3.0
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.27.0
google-auth-oauthlib==1.2.0
greenlet==3.0.3
grpcio==1.60.1
h11==0.14.0
html-parser==0.2
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.20.3
idna==2.10
imageio==2.33.1
importlib-metadata==6.11.0
inquirerpy==0.3.4
IProgress==0.4
ipykernel==6.28.0
ipython==8.20.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.12
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
langchain==0.0.352
langchain-community==0.0.16
langchain-core==0.1.16
langcodes==3.3.0
langsmith==0.0.83
lazy_loader==0.3
llama-index==0.9.39
lxml==5.1.0
Markdown==3.5.2
MarkupSafe==2.1.4
marshmallow==3.20.2
matplotlib==3.8.0
matplotlib-inline==0.1.6
miniful==0.0.6
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.16.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.3
oauthlib==3.2.2
openai==1.2.2
opencv-python==4.7.0.72
opencv-python-headless==4.8.0.74
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.3
pathlib_abc==0.1.1
pathy==0.11.0
pexpect==4.9.0
pfzy==0.3.4
pillow==10.2.0
pinecone-client==3.0.2
platformdirs==4.1.0
plotly==5.18.0
ply==3.11
preshed==3.0.9
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybboxes==0.1.6
pycocotools==2.0.6
pycparser==2.21
pydantic==1.10.9
pydantic_core==2.14.6
pyFUME==0.2.25
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-magic==0.4.27
pytz==2023.3.post1
PyWavelets==1.5.0
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roboflow==1.1.19
rpds-py==0.17.1
rsa==4.9
s3transfer==0.10.0
safetensors==0.4.2
sahi==0.11.15
scikit-image==0.20.0
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
shapely==2.0.2
simpful==2.11.1
simpleeval==0.9.13
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.24
srsly==2.4.8
stack-data==0.6.3
starlette==0.27.0
supervision==0.18.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminado==0.18.0
terminaltables==3.1.10
thinc==8.1.12
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2024.1.30
tiktoken==0.5.2
tinycss2==1.2.1
tokenizers==0.15.1
torch==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.37.1
typer==0.7.0
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.6.0
tzdata==2023.4
ultralytics==8.1.10
uri-template==1.3.0
urllib3==2.0.7
uvicorn==0.23.2
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
yolov5==7.0.13
zipp==3.17.0
